+++ src/index.ts
@@ -9,3 +9,9 @@ export * from "./types/governance_converters";
 export * from "./types/icp";
 export * from "./types/ledger";
 export * from "./utils/neurons.utils";
+export {TxDuplicateError} from './errors/ledger.errors';
+export {TxCreatedInFutureError} from './errors/ledger.errors';
+export {TxTooOldError} from './errors/ledger.errors';
+export {InsufficientFundsError} from './errors/ledger.errors';
+export {InvalidSenderError} from './errors/ledger.errors';
+export {TransferError} from './errors/ledger.errors';
